server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
langgraph==0.1.5
google-generativeai==0.7.2
sqlalchemy==2.0.31
asyncpg==0.29.0
redis==5.0.7
websockets==12.0
structlog==24.2.0
tenacity==8.4.1
